+++ tests/_fstrings.py
@@ -1,0 +1,11 @@
+def test_fstring():
+    assert f"{'hello'}" == "hello"
+
+
+def test_multiline():
+    assert (
+        "world"
+        in f"hello ................................................................ {
+            'world' + '................................................'
+        }"
+    )
