+++ tests/test_fstrings.py
@@ -1,0 +1,1 @@
+from ._fstrings import *  # noqa: F403
